+++ go.mod
@@ -14,5 +14,6 @@ require (
 	github.com/pelletier/go-toml v1.2.0
 	github.com/spf13/afero v1.4.0
 	github.com/spf13/cobra v1.0.0
+	github.com/stretchr/testify v1.6.1
 	github.com/tdewolff/minify/v2 v2.9.5
 )
