+++ testing/bundle/basics.js.out
@@ -1,5 +1,5 @@
 (() => {
-  // testing/bundle/basics.js
+  // bundle/basics.js
   function test() {
     return "test";
   }
